Refactor myRequestHelp page
